+++ src/BlockHorizons/BlockSniper/brush/registration/TypeRegistration.php
@@ -18,6 +18,7 @@ use BlockHorizons\BlockSniper\brush\types\LayerType;
 use BlockHorizons\BlockSniper\brush\types\LeafBlowerType;
 use BlockHorizons\BlockSniper\brush\types\MeltType;
 use BlockHorizons\BlockSniper\brush\types\OverlayType;
+use BlockHorizons\BlockSniper\brush\types\PlantType;
 use BlockHorizons\BlockSniper\brush\types\RegenerateType;
 use BlockHorizons\BlockSniper\brush\types\ReplaceAllType;
 use BlockHorizons\BlockSniper\brush\types\ReplaceTargetType;
@@ -65,6 +66,7 @@ class TypeRegistration{
 		self::registerType(HeatType::class, HeatType::ID);
 		self::registerType(SmoothType::class, SmoothType::ID);
 		self::registerType(ReplaceTargetType::class, ReplaceTargetType::ID);
+		self::registerType(PlantType::class, PlantType::ID);
 	}
 
 	/**